requirements.txt
# Requirements for ESP32 Simulator
paho-mqtt==1.6.1
requests==2.31.0
